refactor: change funciotn of create all format & format word
